refactor: introduce Navigator as a shared navigation channel (#440)
Extract currentScreen + the back stack + navigateTo/navigateToTab/goBack
into a process-scoped Navigator object (mirrors SessionHolder), so feature
ViewModels can drive navigation without routing through SpotifyViewModel.
SpotifyViewModel delegates its nav methods to Navigator and resets it on
construction, preserving "fresh ViewModel == fresh nav" semantics (and
unit-test isolation) while keeping navigation across config changes. Pure
indirection: no screen call sites change and NavigationStackTest passes
unchanged.

Closes #440
